refactor(demo): consolidate cms configuration
