+++ apps/demo/src/app/layout.tsx
@@ -1,43 +1,11 @@
-import { contentPath, initContent } from '@/content'
+import { cms } from '@/cms'
 import { DeepDishProvider } from '@deepdish/cms'
-import { createJsonResolver } from '@deepdish/resolvers/json'
-import { configure } from '@deepdish/ui/config'
-import { typographySchema } from '@deepdish/ui/typography'
 import type { Metadata } from 'next'
 import localFont from 'next/font/local'
-import { z } from 'zod'
+
 import './globals.css'
 
-await initContent()
-
-const blogSchema = z.object({
-  title: z.string(),
-  body: z.string(),
-})
-
-await configure({
-  contracts: {
-    typography: {
-      resolver: createJsonResolver(contentPath, typographySchema, {
-        maxBatchSize: 10,
-      }),
-      schema: typographySchema,
-    },
-    blog: {
-      resolver: createJsonResolver(contentPath, blogSchema, {
-        maxBatchSize: 10,
-      }),
-      schema: blogSchema,
-    },
-  },
-  logging: {
-    enabled: process.env.NODE_ENV === 'development',
-  },
-  settings: {
-    baseUrl: process.env.BASE_URL,
-    draft: process.env.DEEPDISH_MODE === 'draft',
-  },
-})
+await cms()
 
 const geistSans = localFont({
   src: './fonts/GeistVF.woff',
